[Tizen] Reset to 2.3.44 + Revert "Remove WindowAccessible on accessible detroyed"
This reverts commit 25cc45998dae8705da905585f6c67e7a7f5fb408.

+++ dali/internal/graphics/gles-impl/gles-graphics-framebuffer.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Samsung Electronics Co., Ltd.
+ * Copyright (c) 2024 Samsung Electronics Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -179,23 +179,26 @@ bool Framebuffer::InitializeResource()
 
 void Framebuffer::DestroyResource()
 {
-  auto context = mController.GetCurrentContext();
-  auto gl      = mController.GetGL();
-  if(gl && context && mInitialized)
+  if(DALI_LIKELY(!EglGraphicsController::IsShuttingDown()))
   {
-    if(mDepthBufferId)
+    auto context = mController.GetCurrentContext();
+    auto gl      = mController.GetGL();
+    if(gl && context && mInitialized)
     {
-      gl->DeleteRenderbuffers(1, &mDepthBufferId);
-    }
-    if(mStencilBufferId)
-    {
-      gl->DeleteRenderbuffers(1, &mStencilBufferId);
-    }
+      if(mDepthBufferId)
+      {
+        gl->DeleteRenderbuffers(1, &mDepthBufferId);
+      }
+      if(mStencilBufferId)
+      {
+        gl->DeleteRenderbuffers(1, &mStencilBufferId);
+      }
 
-    context->DeleteFramebuffers(1, &mFramebufferId);
+      context->DeleteFramebuffers(1, &mFramebufferId);
 
-    mFramebufferId = 0u;
-    mInitialized   = false;
+      mFramebufferId = 0u;
+      mInitialized   = false;
+    }
   }
 }
